move resources under react-resource-router/resources
