+++ src/controllers/router-store/types.ts
@@ -11,13 +11,12 @@ import {
   Match,
   MatchParams,
   Query,
-  ResourceStoreContext,
-  ResourceStoreData,
   Route,
   RouterContext,
   Routes,
   Plugin,
 } from '../../common/types';
+import type { ResourceStoreContext, ResourceStoreData } from '../../resources';
 
 type PublicStateProperties = {
   action: HistoryAction;
