+++ src/mocks.ts
@@ -7,8 +7,16 @@ export {
   mockRouteContext,
   mockRouteContextProp,
   mockRouterActions,
-  mockRouteResourceResponse,
   mockRouterStoreContext,
   mockRouterStoreContextProp,
   mockRoutes,
 } from './common/mocks';
+
+import { ResourceDependencyError } from './index';
+
+ResourceDependencyError;
+
+/**
+ * @deprecated import from react-resource-router/resources instead
+ */
+export { mockRouteResourceResponse } from './resources';
