+++ src/resources/common/mocks/index.ts
@@ -1,0 +1,9 @@
+export const mockRouteResourceResponse = {
+  loading: false,
+  error: null,
+  data: { foo: 'bar' },
+  promise: null,
+  expiresAt: Date.now(),
+  accessedAt: Date.now(),
+  key: '',
+};
