+++ src/resources/controllers/resource-store/utils/create-loading-slice/index.ts
@@ -1,9 +1,9 @@
-import {
+import type { RouterContext } from '../../../../../index';
+import type {
   ResourceDependencies,
   ResourceStoreContext,
-  RouterContext,
   RouteResource,
-} from '../../../../../common/types';
+} from '../../../../common/types';
 import { GetResourceOptions, PrefetchSlice } from '../../types';
 import { DEFAULT_RESOURCE_MAX_AGE } from '../create-resource/constants';
 import { getExpiresAt } from '../expires-at';
