+++ src/resources/plugin/index.ts
@@ -1,10 +1,9 @@
+import { Plugin, RouterContext } from '../../index';
 import type {
-  Plugin,
   ResourceStoreContext,
-  RouterContext,
   RouteResourceResponse,
   ResourceStoreData,
-} from '../../common/types';
+} from '../common/types';
 import { getResourceStore } from '../controllers/resource-store';
 import { getResourcesForNextLocation } from '../controllers/resource-store/utils';
